refactor(ui): split voice-test-panel into focused modules
427-line panel was doing too many things in one file. Split by concern
so each file has a single, discoverable responsibility:

  voice-test-panel.tsx                (216 lines, was 427)
    Shell: session state + useMutation + <LiveKitRoom> mount.

  voice-test/state.ts                 (36 lines, new)
    WidgetState union, PHASE_LABEL map, AGENT_JOIN_TIMEOUT_MS constant.
    Shared so the shell and RoomController can't drift on state labels.

  voice-test/mic-permission.ts        (57 lines, new)
    ensureMicPermission() — Permissions API → getUserMedia fallback →
    user-friendly error messages. Throws with an actionable string or
    returns silently.

  voice-test/room-controller.tsx      (155 lines, new)
    Child component that runs inside <LiveKitRoom>. Owns the AGENT
    participant wait + 15s timeout, disconnect exposure via ref, and
    the in-call toolbar (mic toggle + hang-up).

No behaviour change — same 268/268 UI tests pass. voice-test-panel.test.tsx
still imports VoiceTestPanel from the same path as before (the public
API of the module is unchanged), so tests didn't need updates.
